4.5 Отрефакторит модуль с утилитами
