+++ src/utils/event.js
@@ -1,0 +1,7 @@
+import dayjs from 'dayjs';
+
+export const humanizeDateToTimeDate = (data) => dayjs(data).format('YYYY-MM-DD[T]HH:mm');
+export const humanizeDateToTimeDateMini = (data) => dayjs(data).format('YYYY-MM-DD');
+export const humanizeDate = (date) => dayjs(date).format('DD/MM/YY HH:mm');
+export const humanizeDateToTime = (date) => dayjs(date).format('HH:mm');
+export const humanizeDateToDayMonth = (data) => dayjs(data).format('MMM DD');
